format test

+++ test/Vault.t.sol
@@ -26,91 +26,49 @@ contract VaultTest is Test, Constants {
         vault.initialize(address(this), address(0x00));
 
         tunnelRouter = new GasPriceTunnelRouter();
-        tunnelRouter.initialize(
-            tssVerifier,
-            vault,
-            address(this),
-            75000,
-            75000,
-            1
-        );
+        tunnelRouter.initialize(tssVerifier, vault, address(this), 75000, 75000, 1);
 
         vault.setTunnelRouter(address(tunnelRouter));
 
         // deploy packet Consumer with specific address.
-        bytes memory packetConsumerArgs = abi.encode(
-            address(tunnelRouter),
-            keccak256("bandchain"),
-            keccak256("eth"),
-            address(this)
-        );
+        bytes memory packetConsumerArgs =
+            abi.encode(address(tunnelRouter), keccak256("bandchain"), keccak256("eth"), address(this));
         address packetConsumerAddr = makeAddr("PacketConsumer");
-        deployCodeTo(
-            "PacketConsumer.sol:PacketConsumer",
-            packetConsumerArgs,
-            packetConsumerAddr
-        );
+        deployCodeTo("PacketConsumer.sol:PacketConsumer", packetConsumerArgs, packetConsumerAddr);
         packetConsumer = PacketConsumer(payable(packetConsumerAddr));
     }
 
     function testDepositWithdrawDeactiveContract() public {
         // deposit
-        uint balanceVaultBefore = address(vault).balance;
+        uint256 balanceVaultBefore = address(vault).balance;
 
         packetConsumer.deposit{value: 0.01 ether}();
 
-        assertEq(
-            vault.balance(packetConsumer.tunnelId(), address(packetConsumer)),
-            0.01 ether
-        );
+        assertEq(vault.balance(packetConsumer.tunnelId(), address(packetConsumer)), 0.01 ether);
         assertEq(address(vault).balance, balanceVaultBefore + 0.01 ether);
-        assertEq(
-            tunnelRouter.isActive(
-                packetConsumer.tunnelId(),
-                address(packetConsumer)
-            ),
-            false
-        );
+        assertEq(tunnelRouter.isActive(packetConsumer.tunnelId(), address(packetConsumer)), false);
 
         // withdraw
         packetConsumer.withdraw(0.01 ether);
-        assertEq(
-            vault.balance(packetConsumer.tunnelId(), address(packetConsumer)),
-            0
-        );
+        assertEq(vault.balance(packetConsumer.tunnelId(), address(packetConsumer)), 0);
         assertEq(address(vault).balance, balanceVaultBefore);
     }
 
     function testDepositWithdrawActiveContract() public {
         // activate + deposit
-        uint balanceVaultBefore = address(vault).balance;
+        uint256 balanceVaultBefore = address(vault).balance;
 
         packetConsumer.activate{value: 0.01 ether}(2);
 
-        assertEq(
-            vault.balance(packetConsumer.tunnelId(), address(packetConsumer)),
-            0.01 ether
-        );
+        assertEq(vault.balance(packetConsumer.tunnelId(), address(packetConsumer)), 0.01 ether);
         assertEq(address(vault).balance, balanceVaultBefore + 0.01 ether);
-        assertEq(
-            tunnelRouter.isActive(
-                packetConsumer.tunnelId(),
-                address(packetConsumer)
-            ),
-            true
-        );
+        assertEq(tunnelRouter.isActive(packetConsumer.tunnelId(), address(packetConsumer)), true);
 
         // withdraw
         vm.expectRevert(IVault.InsufficientRemainingBalance.selector);
         packetConsumer.withdraw(0.01 ether);
 
-        assertEq(
-            tunnelRouter.isActive(
-                packetConsumer.tunnelId(),
-                address(packetConsumer)
-            ),
-            true
-        );
+        assertEq(tunnelRouter.isActive(packetConsumer.tunnelId(), address(packetConsumer)), true);
     }
 
     receive() external payable {}
